Natural Language Processing Of Movie Reviews using NLTK
Natural Language Processing Of Movie Reviews using NLTK. Here in this file, you can get the code for using the python library named as NLTK ( Natural Language ToolKit ), Here I used the NLTK's database for downloading the data, and used the Naive Bayes Classifier of the NLTK for the training of the model and use this model for predicting the results of remaining data which is unseen to the model. And it predicts around 70% of the accuracy. Check out the code file for more information about the use of NLTK library of python and try to play with the code in your desired way.